requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
decorator==4.4.2
Django==3.1.4
django-filter==2.4.0
django-summernote==0.8.11.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
future==0.18.2
geocoder==1.38.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==2.0.0
pylint==2.6.0
pytz==2020.4
ratelim==0.1.6
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1